LU-507 check_write_rcs() return -EPROTO instead of -ENOSPC
remote_rcs[i] is an unsigned value so the check remote_rcs[i] < 0 is
always false.

Reviewed-on: http://review.whamcloud.com/1115
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/osc/osc_request.c
@@ -1197,7 +1197,7 @@ static int check_write_rcs(struct ptlrpc_request *req,
 
         /* return error if any niobuf was in error */
         for (i = 0; i < niocount; i++) {
-                if (remote_rcs[i] < 0)
+                if ((int)remote_rcs[i] < 0)
                         return(remote_rcs[i]);
 
                 if (remote_rcs[i] != 0) {
